Project Structure Refactoring
