--- src/main/java/com/runner/Main.java
@@ -1,13 +1,0 @@
-package com.runner;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
-public class Main {
-
-    private static final Logger logger = LogManager.getLogger(Main.class);
-
-    public static void main(String[] args) {
-        logger.info("LumaProject");
-    }
-}
