+++ src/main/java/selenium/pp/page/AccountPage.java
@@ -1,4 +1,4 @@
-package com.page;
+package selenium.pp.page;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
