+++ src/main/java/selenium/pp/page/AddressBookPage.java
@@ -1,4 +1,4 @@
-package com.page;
+package selenium.pp.page;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
