+++ src/main/java/selenium/pp/page/SignUpPage.java
@@ -1,4 +1,4 @@
-package com.page;
+package selenium.pp.page;
 
 import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
