+++ src/test/java/com/page/SignUpPageTest.java
@@ -1,11 +1,12 @@
 package com.page;
 
-import models.User;
+import api.models.User;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.testng.Assert;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
+import selenium.pp.page.MainPage;
 import runner.BaseTest;
 
 public class SignUpPageTest extends BaseTest {
@@ -23,7 +24,7 @@ public class SignUpPageTest extends BaseTest {
     @Test
     void testSignInUser() {
         final String expectedText = "Thank you for registering with Main Website Store.";
-        logger.info(mainPage.getDriver().getTitle());
+        //logger.info(mainPage.getDriver().getTitle());
 
         User user = new User();
         mainPage.openPage();
@@ -34,7 +35,7 @@ public class SignUpPageTest extends BaseTest {
                 .enterPassword(user.getPassword())
                 .enterConfirmPassword(user.getPassword())
                 .clickCreateAccountButton();
-        logger.info(mainPage.getDriver().getTitle());
+        //logger.info(mainPage.getDriver().getTitle());
         String actualUser = mainPage.getContactInformation();
         Assert.assertTrue(actualUser.contains(user.getFirstname()), "Text does not contain first name: " + user.getFirstname());
         Assert.assertTrue(actualUser.contains(user.getLastname()), "Text does not contain last name: " + user.getLastname());
@@ -44,7 +45,7 @@ public class SignUpPageTest extends BaseTest {
 
         if (mainPage.clickShevron().isLoggedIn()) {
             mainPage.clickLogoutAccount().verifyLogOut();
-            logger.info(mainPage.getSignOutText());
+            //logger.info(mainPage.getSignOutText());
         }
     }
 }
