+++ src/test/java/runner/BaseTest.java
@@ -60,11 +60,11 @@ public abstract class BaseTest {
         }
     }
 
-    protected WebDriver getDriver() {
+    public WebDriver getDriver() {
         return driver;
     }
 
-    protected Page getPage() {
+    public Page getPage() {
         return page;
     }
 }
