+++ src/test/java/selenium/SearchBoxTest.java
@@ -1,3 +1,5 @@
+package selenium;
+
 import org.openqa.selenium.By;
 import org.testng.annotations.Test;
 import runner.BaseTest;
