+++ src/test/java/selpw/SignInExampleTest.java
@@ -1,3 +1,5 @@
+package selpw;
+
 import com.microsoft.playwright.Locator;
 import com.microsoft.playwright.Page;
 import com.microsoft.playwright.options.AriaRole;
